+++ test/hexapdf/type/acro_form/test_form.rb
@@ -574,7 +574,7 @@ describe HexaPDF::Type::AcroForm::Form do
       before do
         @acro_form[:Fields] = [
           @doc.add({T: 'e', Subtype: :Widget, Rect: [0, 0, 0, 1]}),
-          @doc.add({T: 'e', Subtype: :Widget, Rect: [0, 0, 0, 2]}),
+          @merged_field = @doc.add({T: 'e', Subtype: :Widget, Rect: [0, 0, 0, 2]}),
           @doc.add({T: 'Tx2'}),
           @doc.add({T: 'e', Kids: [{Subtype: :Widget, Rect: [0, 0, 0, 3]}]}),
         ]
@@ -585,6 +585,12 @@ describe HexaPDF::Type::AcroForm::Form do
         assert_equal(2, @acro_form.root_fields.size)
         assert_equal([[0, 0, 0, 1], [0, 0, 0, 2], [0, 0, 0, 3]],
                      @acro_form.field_by_name('e').each_widget.map {|w| w[:Rect] })
+      end
+
+      it "deletes the combined and now unneeded field objects" do
+        assert(@acro_form.validate)
+        assert(@merged_field.null?)
+        assert(@doc.object(@merged_field.oid).null?)
       end
     end